refactor: build NamedQuery class family spliting into own module
