+++ pykg2tbl/exceptions.py
@@ -1,0 +1,6 @@
+class WrongInputFormat(Exception):
+    """Raised when this happens..."""
+
+    def __init__(self, message):
+        message = "Wrong input format. QueryResult only allow for List[dicts]"
+        super().__init__(message)
